Hopefully fix mistrusting Anti Virus software
As reported in #120, #119 and multiple times on Discord, the v1.4.5 hotfix was flagged as malware by a few different anti virus softwares. This is likely due to the change in the digital certification the this version had, as I was trying something new it clearly didn't like. Back on having the publisher, and the signing product just be "AssistantComputerControl" like before.